add path str sanitization to create_trial
