+++ smdebug/trials/trial.py
@@ -145,9 +145,10 @@ class Trial(ABC):
             num_times_before_warning -= 1
             if num_times_before_warning < 0:
                 self.logger.warning(
-                    "Waiting to read collections files generated by the training job."
-                    "If this has been a while, you might want to check that the "
-                    "trial is pointed at the right path."
+                    f"Waiting to read collections files generated by the training job,"
+                    f"from {self.path}. "
+                    f"If this has been a while, you might want to check that the "
+                    f"trial is pointed at the right path."
                 )
             else:
                 self.logger.debug(
